pyproject.toml
[project]
name = "beast-mailbox-agent"
version = "0.1.0"
description = "LLM agent that receives and responds to prompts via beast-mailbox-core"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["ai", "llm", "agent", "mailbox", "redis", "messaging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
  "beast-mailbox-core>=0.3.0",
  "openai>=1.0.0",  # For LLM interaction
  "python-dotenv>=1.0.1",
  "tenacity>=8.2.3",
  "typer>=0.12.3",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.0.0",
  "fakeredis>=2.23.0",
  "testcontainers[redis]>=4.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
  "--cov=src/beast_mailbox_agent",
  "--cov-report=xml",
  "--cov-report=term-missing",
  "--verbose",
]

[project.urls]
Homepage = "https://github.com/nkllon/beast-mailbox-agent"
Repository = "https://github.com/nkllon/beast-mailbox-agent"

[project.scripts]
beast-agent = "beast_mailbox_agent.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
